Add Book details page
These changes enable a rendered web page that shows the contents of a
Book model to end users. It is currently bare bones and only does
minimal error handling in terms of accessing and validating. In the
future, among obvious UI changes, more thorough handling and redirecting
of requests for Book details will be implemented.

With these changes, I've taken the liberty of creating a base HTML
template for use throughout the various apps we have. I've not added any
actual items to the UI of the base template since it's irrelevant for
now and is likely work that'd be redone later on anyway as we approach
integration time.

I'm not going to be adding detail pages for Authors, Genres, Publishers,
etc since they are not required at all and are only nice to haves; not
to mention that these kinds of pages would likely be directory pages
that display all work assoc. to that object, which is more in line with
Nico's work on Search/Sort anyway.