Update linting dependency
To improve the functionality of PyLint, I've included the use of a
Django plugin that is much more understanding of the Django environment
which will reduce the amount of false alarms around code checks.

I've also made a quick change around a linter error that was returned
and fixable (whitespace around braces).